Conflicts:
	DropNet.WindowsPhone/DropNet.WindowsPhone.csproj
	DropNet/Client/Client.cs
	DropNet/Client/Files.Async.cs
	DropNet/Client/Files.Sync.cs
	DropNet/Client/User.Async.cs
	DropNet/Client/User.Sync.cs
	DropNet/DropNet.csproj

+++ DropNet/Models/MetaData.cs
@@ -7,7 +7,7 @@ namespace DropNet.Models
     {
         public string Hash { get; set; }
         public bool Thumb_Exists { get; set; }
-        public int Bytes { get; set; }
+        public long Bytes { get; set; }
         public DateTime Modified { get; set; }
         public string Path { get; set; }
         public bool Is_Dir { get; set; }
